🎨 Extract types from services
