+++ src/services/adoption/api.ts
@@ -1,7 +1,10 @@
 import firestore from '@react-native-firebase/firestore';
+
 import animalAPI from '../animal/api';
-import notificationAPI, { NotificationType } from '../notifications/api';
-import { DocumentData, DocumentRefData, DocumentSnapshot } from '../../types/services/firebase';
+import notificationAPI from '../notifications/api';
+
+import { NotificationType } from '../../types/services/Notifications';
+import { DocumentData, DocumentRefData, DocumentSnapshot } from '../../types/services/Firebase';
 import { PaginatedMetaData, filterPaginated } from '../paginated/api';
 
 const addInterestToAnimal = async (animal: DocumentRefData, user: DocumentRefData): Promise<void> => {
