+++ src/types/services/Firebase/index.ts
@@ -7,3 +7,9 @@ export type DocumentSnapshot = FirebaseFirestoreTypes.DocumentSnapshot<DocumentD
 export type Query = FirebaseFirestoreTypes.Query;
 export type QuerySnapshot = FirebaseFirestoreTypes.QuerySnapshot<DocumentData>;
 export type Timestamp = FirebaseFirestoreTypes.Timestamp;
+
+export type QueryParams = {
+  orderBy? : string | undefined,
+  startAfter? : FirebaseFirestoreTypes.DocumentFieldType | undefined,
+  limit? : number | undefined
+}
