+++ src/types/services/Paginated/index.ts
@@ -1,0 +1,7 @@
+export interface PaginatedMetaData {
+  pageNumber: number;
+  pageSize: number;
+  lastElementMarker: unknown[];
+  marker: string;
+  orderByDirection?: 'asc' | 'desc';
+}
